[BUGFIX] Change latitude and longitude data type from Long to Double

+++ src/main/java/com/team25/event/planner/common/dto/LocationResponseDTO.java
@@ -7,6 +7,6 @@ public class LocationResponseDTO {
     private final String country;
     private final String city;
     private final String address;
-    private final Long latitude;
-    private final Long longitude;
+    private final Double latitude;
+    private final Double longitude;
 }
